+++ examples/main/shared/viewer/ViewerPlugins.jsx
@@ -57,6 +57,7 @@ import 'wix-rich-content-plugin-sound-cloud/dist/styles.min.css';
 import 'wix-rich-content-plugin-map/dist/styles.min.css';
 import 'wix-rich-content-plugin-file-upload/dist/styles.min.css';
 import 'wix-rich-content-plugin-giphy/dist/styles.min.css';
+import 'wix-rich-content-text-selection-toolbar/dist/styles.min.css';
 
 import { getBaseUrl } from '../../src/utils';