+++ examples/storybook/stories/Wrapper/Themes.js
@@ -7,7 +7,7 @@ import SegmentedToggle from 'wix-style-react/SegmentedToggle';
 import FormField from 'wix-style-react/FormField';
 import { Layout } from 'wix-style-react/Layout';
 import EditorWrapper from '../Components/EditorWrapper';
-import sourcecode from '!!raw-loader!../Components/EditorWrapperSourceCode';
+import sourcecode from '!!raw-loader!../Components/EditorWrapper';
 
 export default () => {
   return (