+++ CHANGELOG.md
@@ -13,6 +13,10 @@
     Click to see more.
   </summary>
   
+### :bug: Bug Fix
+- `editor-common`
+  - [#647](https://github.com/wix-incubator/rich-content/pull/647) block selection is removed when losing focus
+  
 </details>
 <hr/>